New Watch Page - Stage 5: W-Search
