+++ src/screens/Watch/Components/Search/index.js
@@ -5,10 +5,10 @@ import './index.css';
 import InputBar from './InputBar';
 import Results from './Results';
 
-function SearchWithRedux({ search = SEARCH_INIT }) {
+function SearchWithRedux({ search = SEARCH_INIT, dispatch }) {
   return search.status !== SEARCH_HIDE ? (
     <div id="watch-search-container" className="watch-search">
-      <InputBar search={search} />
+      <InputBar search={search} dispatch={dispatch} />
       <Results search={search} />
     </div>
   ) : null;
